[Refactor] Extract tablet column conversion utilities
