+++ be/src/storage/tablet_column.cpp
@@ -19,73 +19,13 @@
 #include <sstream>
 #include <utility>
 
-#include "storage/metadata_util.h"
-#include "storage/olap_define.h"
 #include "storage/primitive/type_utils.h"
+#include "storage/tablet_column_util.h"
 
 namespace starrocks {
 
 size_t TabletColumn::estimate_field_size(size_t variable_length) const {
     return TypeUtils::estimate_field_size(_type, variable_length);
-}
-
-uint32_t TabletColumn::get_field_length_by_type(LogicalType type, uint32_t string_length) {
-    switch (type) {
-    case TYPE_UNKNOWN:
-    case TYPE_DISCRETE_DOUBLE:
-    case TYPE_STRUCT:
-    case TYPE_MAP:
-    case TYPE_NONE:
-    case TYPE_NULL:
-    case TYPE_FUNCTION:
-    case TYPE_TIME:
-    case TYPE_BINARY:
-    case TYPE_MAX_VALUE:
-    case TYPE_BOOLEAN:
-    case TYPE_TINYINT:
-    case TYPE_UNSIGNED_TINYINT:
-        return 1;
-    case TYPE_SMALLINT:
-    case TYPE_UNSIGNED_SMALLINT:
-        return 2;
-    case TYPE_DATE_V1:
-        return 3;
-    case TYPE_INT:
-    case TYPE_UNSIGNED_INT:
-    case TYPE_FLOAT:
-    case TYPE_DATE:
-    case TYPE_DECIMAL32:
-        return 4;
-    case TYPE_BIGINT:
-    case TYPE_UNSIGNED_BIGINT:
-    case TYPE_DOUBLE:
-    case TYPE_DATETIME_V1:
-    case TYPE_DATETIME:
-    case TYPE_DECIMAL64:
-        return 8;
-    case TYPE_DECIMAL:
-        return 12;
-    case TYPE_LARGEINT:
-    case TYPE_OBJECT:
-    case TYPE_DECIMALV2:
-    case TYPE_DECIMAL128:
-        return 16;
-    case TYPE_DECIMAL256:
-    case TYPE_INT256:
-        return 32;
-    case TYPE_CHAR:
-        return string_length;
-    case TYPE_VARCHAR:
-    case TYPE_HLL:
-    case TYPE_PERCENTILE:
-    case TYPE_JSON:
-    case TYPE_VARIANT:
-    case TYPE_VARBINARY:
-        return string_length + sizeof(get_olap_string_max_length());
-    case TYPE_ARRAY:
-        return string_length;
-    }
-    return 0;
 }
 
 TabletColumn::TabletColumn() = default;
